Graphics Queuing, Render Sleep, Renames
Rendering can now be queued so rendering can be handled (by default) more efficently.  This allows for rendering to be ordered by mesh, by material, and by instance to most effectively load resources.

OpenTK no longer can spin lock 100%, and is now locked to about 10% of what it could previously do.

Materials and meshes now have equality and hash methods so they can be compared correctly.

Program was renamed to EntryPoint, and should be treated as such. Code will be moved out of the class as time goes on.

Code cleanup